Separate tests and implementation
